add mac os support

+++ log.c
@@ -37,7 +37,7 @@ static char *print_time()
     nowtm = localtime(&now);
     char timebuf[32];
     strftime(timebuf, 32, "%Y-%m-%d %H:%M:%S", nowtm);
-    snprintf(curr_time, 32, "%s %06ld",timebuf, tv.tv_usec);
+    snprintf(curr_time, 32, "%s %06ld",timebuf, (long)tv.tv_usec);
     return curr_time;
 }
 static int log_open_file()
